Refactor (APIMF-3491): extracted object property parsing
